+++ src/components/Login/LoginButton.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, Text, useDisclosure } from '@chakra-ui/react';
 import Button from 'components/common/Button';
-import BottomDrawer from 'components/Login/BottonDrawer';
+import BottomDrawer from 'components/Login/BottomDrawer';
 import GoogleButton from 'components/Login/OAuth/GoogleButton';
 import KakaoButton from 'components/Login/OAuth/KakaoButton';
 import Image from 'next/image';
